ui_path and alternative route
