+++ AppVoor/resources/ui_path.py
@@ -1,0 +1,33 @@
+ui_window = {"home": ".\\resources\\forms\\QT_Voorspelling_Home.ui",
+             "dataset": ".\\resources\\forms\\QT_Voorspelling_DataSet.ui",
+             "model": ".\\resources\\forms\\QT_Voorspelling_Modelo.ui",
+             "prediction_type": ".\\resources\\forms\\QT_Voorspelling_TipoPrediccion.ui",
+             "classification": ".\\resources\\forms\\QT_Voorspelling_TipoP_Clasificacion.ui",
+             "regression": ".\\resources\\forms\\QT_Voorspelling_TipoP_Regresion.ui",
+             "clustering": ".\\resources\\forms\\QT_Voorspelling_TipoP_Agrupamiento.ui",
+             "feature_selection": ".\\resources\\forms\\QT_Voorspelling_Caracteristicas.ui",
+             "feature_selection_method": ".\\resources\\forms\\QT_Voorspelling_CaracteristicasMetodo.ui",
+             "hyperparameter_search": ".\\resources\\forms\\QT_Voorspelling_Hiperparametros.ui",
+             "hyperparameter_search_method": ".\\resources\\forms\\QT_Voorspelling_HiperparametrosMetodo.ui",
+             "result_screen": ".\\resources\\forms\\QT_Voorspelling_Resultado.ui",
+             "result_final": ".\\resources\\forms\\QT_Voorspelling_ResultadoFinal.ui",
+             "LinearSVC": ".\\resources\\forms\\QT_Voorspelling_ByHand_LinearSVC.ui",
+             "SVC": ".\\resources\\forms\\QT_Voorspelling_ByHand_SVC_rbf.ui",
+             "KNeighborsClassifier": ".\\resources\\forms\\QT_Voorspelling_ByHand_KNN.ui",
+             "GaussianNB": ".\\resources\\forms\\QT_Voorspelling_ByHand_GaussianNaiveBayes.ui",
+             "LinearSVR": ".\\resources\\forms\\QT_Voorspelling_ByHand_LinearSVR.ui",
+             "SVR": ".\\resources\\forms\\QT_Voorspelling_ByHand_SVR_rbf.ui",
+             "Lasso": ".\\resources\\forms\\QT_Voorspelling_ByHand_Lasso.ui",
+             "SGDClassifier": ".\\resources\\forms\\QT_Voorspelling_ByHand_SGD.ui",
+             "AffinityPropagation": ".\\resources\\forms\\QT_Voorspelling_ByHand_AffinityPropagation.ui",
+             "KMeans": ".\\resources\\forms\\QT_Voorspelling_ByHand_KMeans.ui",
+             "MiniBatchKMeans": ".\\resources\\forms\\QT_Voorspelling_ByHand_MiniBatchKMeans.ui",
+             "MeanShift": ".\\resources\\forms\\QT_Voorspelling_ByHand_MeanShift.ui"
+             }
+
+ui_icons = {"16px": ('.\\resources\\icos\\voorspelling_logo_ico_16px.ico', 16),
+            "32px": ('.\\resources\\icos\\voorspelling_logo_ico_32x.ico', 32),
+            "48px": ('.\\resources\\icos\\voorspelling_logo_ico_48px.ico', 48),
+            "54px": ('.\\resources\\icos\\voorspelling_logo_ico_54px.ico', 54),
+            "256px": ('.\\resources\\icos\\voorspelling_logo_ico_256px.ico', 256),
+            }
